Refactor CopyInput out into it's own component
